fix débito técnico - mínimo possível

+++ backend/src/main/java/com/vinculohub/backend/repository/ProjectRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface ProjectRepository extends JpaRepository<Project, Long> {
-    List<Project> findAllByNpoId(Integer npoId);
+    List<Project> findAllByNpoId(Long npoId);
 }
